requirements.txt
Django==4.2.0
dadata==21.10.1
djangorestframework==3.14.0
django-filter==23.2
